# perf(ext/web): flatten op arguments for text_encoding