Towards #14893 
 
 1.4x improvement in CSV parsing.

```
# Before

time deno run --allow-env --allow-read --allow-hrtime deno-csv-CSVReader.ts 500000-Records.csv
Read 500001 lines for 19.435 seconds

________________________________________________________
Executed in   19.50 secs    fish           external
   usr time   17.47 secs    9.85 millis   17.46 secs
   sys time    3.55 secs    2.47 millis    3.54 secs
```

```
# After

time ../../deno/target/release/deno run --allow-env --allow-read --allow-hrtime deno-csv-CSVReader.ts 500000-Records.csv
Read 500001 lines for 14.889 seconds

________________________________________________________
Executed in   14.99 secs    fish           external
   usr time   13.13 secs    7.14 millis   13.13 secs
   sys time    3.18 secs    2.50 millis    3.18 secs
```